+++ src/js/index.js
@@ -23,8 +23,8 @@ import labs from "./labs.js";
 import wind from "./wind.js";
 import s2s from "./s2s.js";
 import * as config from "./config.js";
-import * as places from "./places.js";
-import * as zooms from "./zooms.js";
+import * as places from "../data/places.js";
+import * as zooms from "../data/zooms.js";
 import * as translate from "./translate.js";
 import * as no2data from "../data/no2.json";
 import * as stations from "../data/stations.json";
